Enable scrollback
Apply st-scrollback-20200419-72e3f6c.diff
Apply st-scrollback-mouse-20191024-a2c479c.diff
Apply st-scrollback-mouse-altscreen-20200416-5703aa0.diff

These add a line history.
The third patch lets the mouse only scroll through this history when not
in alternate screen. When in alternate screen (like less and vim for
example) ^H and ^E are sent so that the content of that program is
scrolled.

Set bindings for scrolling and document them in the README.
Increase default history size.