Introduce SEL_LINES to mouse-select linewise
Use sel.snap to implement a new selection type SEL_LINES. With it the
mouse selection can be done linewise, similar to vim's linewise visual
mode. Also add a default mask for it to selmasks.

Add a hint to forcemousemod above the definition of selmasks.

Introduce SNAP_NONE to the selection_snap enum that encodes 0. Change
all hard-coded uses of it.